Здесь сконфигурирован каталог. Много изменений из-за скрипта по рефакторингу кода

+++ src/Business/Services/FileService.cs
@@ -1,10 +1,10 @@
+using AutoMapper;
 using Forpost.Business.Abstract.Services;
+using Forpost.Business.Models.Files;
+using Forpost.Common;
 using Forpost.Store.Entities;
 using Forpost.Store.Repositories.Abstract.Repositories;
 using Microsoft.Extensions.Configuration;
-using AutoMapper;
-using Forpost.Business.Models.Files;
-using Forpost.Common;
 
 namespace Forpost.Business.Services;
 
@@ -25,21 +25,18 @@ internal sealed class FileService : IFileService
     {
         // ParentId - это любой ID из БД, к которому нужно привязать файл.
         // Путь к файлу Files/ParentId/FileName
-        var filePath = Path.Combine(model.ParentId.ToString(), $"{(model.FileName)}");
-            
+        var filePath = Path.Combine(model.ParentId.ToString(), $"{model.FileName}");
+
         var fullPath = Path.Combine(_uploadFilePath, filePath);
-            
+
         var directory = Path.GetDirectoryName(fullPath);
-            
-        if (Directory.Exists(directory) is false)
-        {
-            Directory.CreateDirectory(directory);
-        }
+
+        if (Directory.Exists(directory) is false) Directory.CreateDirectory(directory);
 
         await File.WriteAllBytesAsync(fullPath, model.Content);
         var fileEntity = _mapper.Map<FileEntity>(model);
         fileEntity.FilePath = fullPath;
-            
+
         return await _filesRepository.AddAsync(fileEntity, cancellationToken);
     }
 
@@ -47,10 +44,10 @@ internal sealed class FileService : IFileService
     {
         var file = await _filesRepository.GetByIdAsync(id, cancellationToken);
         file.EnsureFoundBy(x => x.Id, id);
-            
+
         var fullPath = Path.Combine(file.FilePath);
         var fileContent = await File.ReadAllBytesAsync(fullPath);
-            
+
         var downloadFile = _mapper.Map<DownloadFileModel>(file);
         downloadFile.FileContent = fileContent;
         return downloadFile;
@@ -64,10 +61,7 @@ internal sealed class FileService : IFileService
         var fullPath = Path.Combine(file.FilePath);
         await _filesRepository.DeleteByIdAsync(id, cancellationToken);
 
-        if (File.Exists(fullPath))
-        {
-            File.Delete(fullPath);
-        }
+        if (File.Exists(fullPath)) File.Delete(fullPath);
     }
 
     public async Task<IReadOnlyList<FileEntity>> GetAllFilesAsync(Guid parentId, CancellationToken cancellationToken)
